+++ libgfx/include/detail/texture.hh
@@ -87,10 +87,10 @@ public:
     explicit TextureRenderer(gfx::Window& window);
 
     void draw(
-        int x,
-        int y,
-        int width,
-        int height,
+        float x,
+        float y,
+        float width,
+        float height,
         const gfx::IRotation& rotation,
         const gfx::Texture& texture,
         glm::mat4 view
