+++ libgfx/src/line.cc
@@ -47,7 +47,7 @@ LineRenderer::LineRenderer(gfx::Window& window)
     glBindBuffer(GL_ARRAY_BUFFER, 0);
 }
 
-void LineRenderer::draw(int x0, int y0, int x1, int y1, gfx::Color color, glm::mat4 view) {
+void LineRenderer::draw(float x0, float y0, float x1, float y1, gfx::Color color, glm::mat4 view) {
 
     m_vertices.push_back({ x0, y0 });
     m_vertices.push_back({ x1, y1 });
