+++ libgfx/src/rectangle.cc
@@ -50,10 +50,10 @@ RectangleRenderer::RectangleRenderer(gfx::Window& window)
 }
 
 void RectangleRenderer::draw(
-    int x,
-    int y,
-    int width,
-    int height,
+    float x,
+    float y,
+    float width,
+    float height,
     const gfx::IRotation& rotation,
     gfx::Color color,
     glm::mat4 view
